move: auditing test 패키지 경로 이동
