+++ src/main/java/com/caffeine/gwanghwamun/domain/test/repository/TestRepository.java
@@ -1,0 +1,6 @@
+package com.caffeine.gwanghwamun.domain.test.repository;
+
+import com.caffeine.gwanghwamun.domain.test.entity.TestEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TestRepository extends JpaRepository<TestEntity, Long> {}
